requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Jinja2
PyYAML
discord.py
python_dateutil
pytz
